requirements.txt
fastapi
uvicorn[standard]
python-multipart
openpyxl
python-dotenv
